Struct trillium_server_common::Config

pub struct Config<ServerType, AcceptorType> { /* fields omitted */ }

Primary entrypoint for configuring and running a trillium server

The associated methods on this struct are intended to be chained.

Example

// in reality, you'd use trillium_smol_server, trillium_async_std_server, trillium_tokio_server, etc
trillium_testing::server::config()
    .with_port(8080) // the default
    .with_host("localhost") // the default
    .with_nodelay()
    .without_signals()
    .run(|conn: trillium::Conn| async move { conn.ok("hello") });

In order to use this to implement a trillium server, see [trillium_server_common::ConfigExt]

Implementations

impl<ServerType, AcceptorType> Config<ServerType, AcceptorType> where
ServerType: Server,
AcceptorType: Acceptor<ServerType::Transport>,

pub fn run<H: Handler>(self, h: H)

Starts an async runtime and runs the provided handler with this config in that runtime. This is the appropriate entrypoint for applications that do not need to spawn tasks outside of trillium's web server. For applications that embed a trillium server inside of an already-running async runtime, use Config::run_async

pub async fn run_async(self, handler: impl Handler)

Runs the provided handler with this config, in an already-running runtime. This is the appropriate entrypoint for an application that needs to spawn async tasks that are unrelated to the trillium application. If you do not need to spawn other tasks, Config::run is the preferred entrypoint

pub fn with_port(self, port: u16) -> Self

Configures the server to listen on this port. The default is the PORT environment variable or 8080

pub fn with_host(self, host: &str) -> Self

Configures the server to listen on this host or ip address. The default is the HOST environment variable or "localhost"

pub fn without_signals(self) -> Self

Configures the server to NOT register for graceful-shutdown signals with the operating system. Default behavior is for the server to listen for SIGINT and SIGTERM and perform a graceful shutdown.

pub fn with_nodelay(self) -> Self

Configures the tcp listener to use TCP_NODELAY. See https://en.wikipedia.org/wiki/Nagle%27s_algorithm for more information on this setting.

pub fn with_acceptor<A: Acceptor<ServerType::Transport>>(
self,
acceptor: A
) -> Config<ServerType, A>

Configures the tls acceptor for this server

impl<ServerType> Config<ServerType, ()>

pub fn new() -> Self