Trait trillium_server_common::Connector

pub trait Connector: Send + Sync + 'static {
    type Transport: Transport;

    fn connect<'life0, 'life1, 'async_trait>(
        &'life0 self,
        url: &'life1 Url
    ) -> Pin<Box<dyn Future<Output = Result<Self::Transport>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn spawn<Fut: Future<Output = ()> + Send + 'static>(&self, fut: Fut);
}

Interface for runtime and tls adapters for the trillium client

See trillium_client for more information on usage.

Required Associated Types

type Transport: Transport

Required Methods

fn connect<'life0, 'life1, 'async_trait>(
&'life0 self,
url: &'life1 Url
) -> Pin<Box<dyn Future<Output = Result<Self::Transport>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Initiate a connection to the provided url, using the configuration.

Async trait signature:

async fn connect(url: &Url, config: &Self::Config) -> std::io::Result<Self::Transport>;

fn spawn<Fut: Future<Output = ()> + Send + 'static>(&self, fut: Fut)